@@ -1,6 +1,6 @@
 import { createStore, applyMiddleware } from 'redux';
 import logger from 'redux-logger';
-import thunk from 'redux-thunk';
+import { thunk } from 'redux-thunk';
 import allReducers from './reducers';
 
 const allMiddlewares = [thunk];
